requirements.txt
pyttsx3==2.90
Pillow==8.4.0
speechrecognition==3.8.1
tk==0.1.0
num2words==0.5.10
pywhatkit==6.1
keyboard==0.13.5
speedtest-cli==2.1.3
